Some simple enhancements to the approximation.
